+++ studio/web/web.go
@@ -4,5 +4,5 @@ package web
 
 import "embed"
 
-//go:embed index.html app.js style.css charset.js
+//go:embed index.html app.js style.css charset.js altcharset.js
 var FS embed.FS